progress

+++ server/src/main/java/com/supersidor/flightmap/model/PlanePosition.kt
@@ -1,3 +1,3 @@
 package com.supersidor.flightmap.model
 
-data class PlanePosition (val userId:Int,val heading:Int,val longitude:Double,val latitude:Double)
+data class PlanePosition (val userId:Long,val heading:Int,val longitude:Double,val latitude:Double)
